requirements.txt
Django==3.2.22
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0 